requirements-test.txt
requests
psycopg2-binary
PyJWT==2.*
cryptography
bleach==5.0.1
lxml==4.9.2
pandas==2.0.1
numpy==1.24.2
html5lib==1.1
beautifulsoup4==4.12.3
pytest
boto3